Hausübung/Übung_1/IIB2-Demo/IIB2-Demo/requirements.txt
Click==7.0

Flask==1.0.2

itsdangerous==1.1.0

Jinja2==2.10.1

MarkupSafe==1.1.1

mysql-connector==2.2.9

pip==10.0.1

setuptools==39.0.1

Werkzeug==0.15.2